Reorg and make Time Lapse Polygons validation work
