+++ transit-network-analysis-tools/Transit Network Analysis Tools.pyt.xml
@@ -1,2 +1,2 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<metadata xml:lang="en"><Esri><CreaDate>20190516</CreaDate><CreaTime>10340500</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce><ModDate>20190516</ModDate><ModTime>105333</ModTime></Esri><toolbox name="Transit Network Analysis Tools" alias="TransitNetworkAnalysisTools"><arcToolboxHelpPath>c:\program files (x86)\arcgis\desktop10.7\Help\gp</arcToolboxHelpPath><toolsets/></toolbox><dataIdInfo><idCitation><resTitle>Transit Network Analysis Tools</resTitle></idCitation></dataIdInfo><distInfo><distributor><distorFormat><formatName>ArcToolbox Toolbox</formatName></distorFormat></distributor></distInfo></metadata>
+<metadata xml:lang="en"><Esri><CreaDate>20190516</CreaDate><CreaTime>10340500</CreaTime><ArcGISFormat>1.0</ArcGISFormat><SyncOnce>TRUE</SyncOnce><ModDate>20190516</ModDate><ModTime>120915</ModTime></Esri><toolbox name="Transit Network Analysis Tools" alias="TransitNetworkAnalysisTools"><arcToolboxHelpPath>c:\program files (x86)\arcgis\desktop10.7\Help\gp</arcToolboxHelpPath><toolsets/></toolbox><dataIdInfo><idCitation><resTitle>Transit Network Analysis Tools</resTitle></idCitation></dataIdInfo><distInfo><distributor><distorFormat><formatName>ArcToolbox Toolbox</formatName></distorFormat></distributor></distInfo></metadata>
